fix some promblems in photos contorollers and add user order managment {add - delete - show one - show all}

+++ app/Http/Controllers/PhotoController.php
@@ -45,18 +45,22 @@ $image=null;
 $filename=null;
 $save_path=storage_path('\uploads');
 $photo_id_in_user_table = User::find($user_id)->photo_id ;
+$photo_name_in_DB=null;
 $deleted='no path in the DB';
 
 // there we delete the old image from the Storage file
 //if the user send the photo then we store the new
 //if the user dont send a photo then he dont want a profile photo so he want to
 //delete the old image if it exist
+if ($photo_id_in_user_table!=null) {
+    $photo_name_in_DB=photo::find($photo_id_in_user_table)->path;
+    if($photo_name_in_DB !=null){
+        $deleted= File::delete($save_path.'\\' .$photo_name_in_DB);
 
-$photo_name_in_DB=photo::find($photo_id_in_user_table)->path;
-if($photo_name_in_DB !=null){
-    $deleted= File::delete($save_path.'\\' .$photo_name_in_DB);
-
+    }
 }
+
+
 
             if($request->hasFile('photo')){
               $image = $request->file('photo');
@@ -86,7 +90,7 @@ if($photo_name_in_DB !=null){
 
 
 
-        return response()->json(['new photo'=>$new_photo[0] ,'url' =>$save_path.'\\'.$filename ,'is created'=>$is_created ,'photo id in the company table'=>User::find($user_id)->photo_id  ,'photo in DB'=>$photo_name_in_DB ,'deleted'=>$deleted ,'delete path'=>$save_path.'\\' .$photo_name_in_DB]);
+        return response()->json(['new photo'=>$new_photo[0] ,'url' =>$save_path.'\\'.$filename ,'is created in database'=>$is_created ,'photo id in the users table'=>User::find($user_id)->photo_id  ,'photo deleted in DB'=>$photo_name_in_DB ,'deleted from storage file'=>$deleted ,'delete path'=>$save_path.'\\' .$photo_name_in_DB]);
     }
